chiquito_financiero/app/requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
openpyxl>=3.1.2
plotly>=5.20.0
fpdf2>=2.7.9
python-dotenv>=1.0.0
watchdog>=4.0.0
